+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotExceptionImpl.java
@@ -455,7 +455,6 @@ final class PolyglotExceptionImpl extends AbstractExceptionImpl {
             if (!languageContext.isCreated()) {
                 return null;
             }
-
             return languageContext.getInternalFileSystemContext();
         }
     }